breaking: offline store can be initialized with custom serializers
