+++ packages/offix-client/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./OfflineClient";
-export * from "./ApolloOfflineClient";
+export * from "./apollo/ApolloOfflineClient";
 export * from "offix-offline";
 export * from "offix-cache";
